Renamed gauge force to gauge test in general cases as appropriate
