+++ include/gauge_path_quda.h
