+++ tests/CMakeLists.txt
@@ -200,10 +200,10 @@ if(QUDA_DIRAC_STAGGERED)
   install(TARGETS hisq_unitarize_force_test ${QUDA_EXCLUDE_FROM_INSTALL} DESTINATION ${CMAKE_INSTALL_BINDIR})
 endif()
 
-add_executable(gauge_force_test gauge_force_test.cpp)
-target_link_libraries(gauge_force_test ${TEST_LIBS})
-quda_checkbuildtest(gauge_force_test QUDA_BUILD_ALL_TESTS)
-install(TARGETS gauge_force_test ${QUDA_EXCLUDE_FROM_INSTALL} DESTINATION ${CMAKE_INSTALL_BINDIR})
+add_executable(gauge_path_test gauge_path_test.cpp)
+target_link_libraries(gauge_path_test ${TEST_LIBS})
+quda_checkbuildtest(gauge_path_test QUDA_BUILD_ALL_TESTS)
+install(TARGETS gauge_path_test ${QUDA_EXCLUDE_FROM_INSTALL} DESTINATION ${CMAKE_INSTALL_BINDIR})
 
 add_executable(gauge_alg_test gauge_alg_test.cpp)
 target_link_libraries(gauge_alg_test ${TEST_LIBS})
@@ -1181,10 +1181,10 @@ endforeach(prec)
   
 foreach(prec IN LISTS TEST_PRECS)
 
-  add_test(NAME gauge_force_${prec}
-           COMMAND ${QUDA_CTEST_LAUNCH} $<TARGET_FILE:gauge_force_test> ${MPIEXEC_POSTFLAGS}
+  add_test(NAME gauge_path_${prec}
+           COMMAND ${QUDA_CTEST_LAUNCH} $<TARGET_FILE:gauge_path_test> ${MPIEXEC_POSTFLAGS}
                    --dim 2 4 6 8 --prec ${prec}
-                   --gtest_output=xml:gauge_force_test_${prec}.xml)
+                   --gtest_output=xml:gauge_path_test_${prec}.xml)
 
   if(QUDA_DIRAC_STAGGERED)
     add_test(NAME unitarize_link_${prec}
